Fix: Extract class definitions, indentation
